Major refactor of models and parsing
Parsing is now based on the multi-languages page of the episode to
prevent random behaviour when no subtitle is found for a given language.

Parsing has been refactored for code clarity (more readable, more SRP),
it is now closer to the webpage DOM hierarchy, it better supports HI
(not only the absence of HI) and it adds support for subtitles marked
as corrected.

The `Search` object has been removed in favour of a more intuitive
class hierarchy.

This is a major breakthrough on the road to 4.0 :tada: